Extract PHP Validator from main project
Xulieta is evolving to a composable plugin system on version 1.0.x.
So the [main package][xulieta] have the foundation to read common documents and
the logic to make it extensible by separated packages. This way each project
can choose exactly what should be checked and lower the barrier have more
people comming up with new validators.

[xulieta]: https://github.com/codelicia/xulieta

Signed-off-by: 💻 Eher <alexandre@eher.com.br>